Add sidebar buttons for running CodeChecker log
* Add console preview option to run log
